refactor(vault-index): split tfidf.py + themes.py out of vault_index.py (#229) (#232)
* refactor(vault-index): split tfidf.py + themes.py out of vault_index.py (#229)

Behavior-neutral extraction (Slice A of #53). Acyclic graph tfidf<-themes<-
vault_index; themes lazy-imports _connect to break the cycle; explicit named
re-export shim keeps vault_index.<symbol> resolving. _NEGATION_TERMS placed in
themes.py (corrects parent spec). Existing test_tfidf/themes/surprise pass
unmodified; adds test_module_split_parity.py.


* docs(architecture): add tfidf.py + themes.py modules (#229)


* docs(changelog): note vault_index.py module split (#229)


* test(vault-index): pin lazy _connect seam to canonical bare vault_index (#229)

Phase-1 deep-review follow-up. Adds test_themes_lazy_connect_resolves_to_vault_index
asserting the lazy `from vault_index import _connect` in assign_to_theme binds the
canonical bare vault_index module (the form every runtime caller uses), making the
cycle-break seam explicit and fast-failing. Expands the lazy-import comment to
document the bare-form contract and its monkeypatch implication. No behavior change.


* docs(architecture): fix test count + term_df attribution post-split (#229)

Phase-2 adversarial-review survivors: bump unit-test fileCount 64->65 (new
test_module_split_parity.py); move the term_df ON CONFLICT clause from the
themes component to tfidf (where _update_term_df now lives).


---------
