+++ docs/architecture/architecture.json
@@ -3,7 +3,7 @@
   "displayName": "Obsidian Brain",
   "version": "2.7.1",
   "tagline": "A Claude Code plugin that turns an Obsidian vault into a persistent brain across sessions — auto-logging, curated knowledge capture, snapshot-on-compact, and fast cross-project recall, all via direct filesystem writes.",
-  "lastUpdated": "2026-06-12",
+  "lastUpdated": "2026-06-14",
   "repository": "https://github.com/abhattacherjee/obsidian-brain",
   "defaultBranch": "main",
   "techStack": {
@@ -15,7 +15,7 @@
     "realtime": { "protocol": "none", "endpoint": "direct filesystem writes — no MCP server, no REST API, no daemon" },
     "ai": { "provider": "Claude CLI (claude -p)", "model": "haiku default; escalates haiku to sonnet to opus on empty output (#165)", "notes": "summarization deferred from hooks to /recall" },
     "vectorSearch": { "engine": "SQLite FTS5 (BM25) + 7-signal rerank; TF-IDF cosine themes with ACT-R activation and Free-Energy surprise" },
-    "tests": { "unit": { "runner": "pytest", "count": 64, "environments": ["python3.12"], "coverage": { "lines": 90 } } },
+    "tests": { "unit": { "runner": "pytest", "count": 65, "environments": ["python3.12"], "coverage": { "lines": 90 } } },
     "packageManager": "none (stdlib only)"
   },
   "diagram": {
@@ -87,7 +87,8 @@
         { "id": "search-vault", "name": "search_vault()", "kind": "function", "files": ["hooks/vault_index.py"], "purpose": "Sanitizes the query to AND-mode FTS5 (hyphens to spaces), runs bm25 with title 10x and tags 5x weighting, OR-fallback when AND is empty, reranks, batch-logs access, strips body before returning." },
         { "id": "rerank", "name": "rerank_results() — 7-signal scorer", "kind": "function", "files": ["hooks/vault_index.py"], "purpose": "Weighted sum: proximity 0.25, bm25 0.20, activation 0.20, type 0.10, recency 0.10, importance 0.10, density 0.05. Type score is context-adaptive (debugging/standup/emerge/search/general)." },
         { "id": "access-log", "name": "log_access() / batch_activations()", "kind": "datastore-client", "files": ["hooks/vault_index.py"], "purpose": "Records access events (cascading snapshot accesses to the parent session) and computes ACT-R base-level activation, reserving 0.0 for notes with no history." },
-        { "id": "themes", "name": "assign_to_theme() / detect_surprise()", "kind": "module", "files": ["hooks/vault_index.py"], "purpose": "TF-IDF (smoothed IDF, sparse top-50) cosine clustering into running-average centroids (threshold 0.3); Free-Energy surprise = negation-proximity over top shared terms. term_df maintained via ON CONFLICT DO UPDATE, never blind REPLACE." },
+        { "id": "tfidf", "name": "tfidf.py — TF-IDF primitives (leaf)", "kind": "module", "files": ["hooks/tfidf.py"], "purpose": "Stdlib-only leaf (no vault_index or themes imports). Exports _tokenize_for_tfidf, _compute_tfidf_vector, _cosine_similarity, _update_term_df, _STOPWORDS, _TOKEN_RE. Smoothed IDF, sparse top-50 vectors, cosine similarity. Imported by themes.py and re-exported from vault_index.py for back-compat; kept acyclic by design. term_df maintained via ON CONFLICT DO UPDATE, never blind REPLACE." },
+        { "id": "themes", "name": "themes.py — assign_to_theme() / detect_surprise()", "kind": "module", "files": ["hooks/themes.py"], "purpose": "Theme assignment + Free-Energy surprise detection, extracted from vault_index.py (#229). Depends on tfidf (one-directional); lazy-imports vault_index._connect at call-time to break the import cycle. TF-IDF cosine clustering into running-average centroids (threshold 0.3); surprise = negation-proximity over top shared terms." },
         { "id": "rebuild-index", "name": "rebuild_index() — Friston-preserving", "kind": "function", "files": ["hooks/vault_index.py"], "purpose": "Default non-destructive rebuild preserves access_log, themes and theme_members (activation/centroid/surprise), with an absolute-path safety guard that skips orphan-prune on path-format drift. full=True is the legacy total wipe for corrupt schema." },
         { "id": "upsert-note", "name": "_upsert_note() — DELETE-then-INSERT", "kind": "function", "files": ["hooks/vault_index.py"], "purpose": "Deliberately avoids INSERT OR REPLACE: explicit DELETE+INSERT preserving the prior importance value, and issuing the contentless-FTS5 delete op with old column values so the FTS index never accumulates orphans." },
         { "id": "vault-stats-compute", "name": "compute_stats()", "kind": "function", "files": ["hooks/vault_stats.py"], "purpose": "Read-only aggregation for /vault-stats: signal coverage (per-note activation + importance), access-by-context, top-accessed, importance distribution, and snapshot integrity metrics (orphans, broken backlinks, summarized fraction)." }
@@ -378,7 +379,7 @@
     ]
   },
   "testing": {
-    "unitSuite": { "tool": "pytest", "fileCount": 64, "duration": "fast (stdlib, no network)", "environments": { "python": "3.12" }, "coverage": { "lines": 90, "source": "hooks", "omits": ["obsidian_utils.py (I/O monolith)", "emerge_cli.py", "deep_cli.py"] } },
+    "unitSuite": { "tool": "pytest", "fileCount": 65, "duration": "fast (stdlib, no network)", "environments": { "python": "3.12" }, "coverage": { "lines": 90, "source": "hooks", "omits": ["obsidian_utils.py (I/O monolith)", "emerge_cli.py", "deep_cli.py"] } },
     "manualMatrix": { "doc": "CLAUDE.md (Development Commands)", "devices": ["hook simulation: pipe synthetic SessionEnd/SessionStart/PreCompact JSON into installed hooks", "json.load validation of hooks.json and plugin.json"], "rationale": "No build/test framework for the prompt-based skills; hooks are validated by feeding synthetic lifecycle JSON and asserting vault writes + telemetry outcomes." }
   },
   "scripts": [
